requirements.txt
Flask>=2.0.1
google-cloud-storage>=2.3.0
rasterio
shapely
python-dotenv
pyproj
